refactor: clarify settings organization
