+++ messages/ar.json
@@ -83,7 +83,7 @@
   "settingsTitle": "الإعدادات",
   "settingsOpen": "فتح الإعدادات",
   "settingsClose": "إغلاق الإعدادات",
-  "settingsSectionAppearance": "المظهر",
+  "settingsSectionGeneral": "عام",
   "settingsSectionSwitchAccess": "الوصول باستخدام المفتاح",
   "switchScanningEnabled": "تفعيل المسح بالمفتاح",
   "switchScanningMethod": "طريقة المسح",
@@ -113,11 +113,11 @@
   "speechVoice": "الصوت",
   "speechRate": "السرعة",
   "speechPitch": "حدّة الصوت",
-  "speechVolume": "مستوى الصوت",
+  "speechVolume": "مستوى صوت الكلام",
   "speechPreview": "استماع",
   "speechVoicePreview": "مرحبًا، هذا هو صوتي!",
-  "playbackHighlight": "تمييز الكلمات أثناء التحدث",
-  "tileSaturation": "شدة الألوان",
+  "playbackHighlight": "تمييز الكلمات المنطوقة",
+  "tileSaturation": "شدة ألوان المربعات",
   "tileBorders": "إظهار حدود الرموز",
   "aiCustomInstructions": "تعليمات مخصصة",
   "aiCustomInstructionsPlaceholder": "مثل: تحدث بنبرة ساخرة",
