+++ messages/bg.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Настройки",
   "settingsOpen": "Отваряне на настройките",
   "settingsClose": "Затваряне на настройките",
-  "settingsSectionAppearance": "Външен вид",
+  "settingsSectionGeneral": "Общи",
   "settingsSectionSwitchAccess": "Достъп с превключвател",
   "switchScanningEnabled": "Включване на сканирането с превключвател",
   "switchScanningMethod": "Метод на сканиране",
@@ -113,11 +113,11 @@
   "speechVoice": "Глас",
   "speechRate": "Скорост",
   "speechPitch": "Височина",
-  "speechVolume": "Сила на звука",
+  "speechVolume": "Сила на звука на речта",
   "speechPreview": "Прослушване",
   "speechVoicePreview": "Здравей, това е моят глас!",
-  "playbackHighlight": "Осветяване на думите по време на говорене",
-  "tileSaturation": "Интензивност на цвета",
+  "playbackHighlight": "Осветяване на изговорените думи",
+  "tileSaturation": "Интензивност на цвета на плочките",
   "tileBorders": "Показване на рамки на плочките",
   "aiCustomInstructions": "Персонализирани инструкции",
   "aiCustomInstructionsPlaceholder": "напр. говори със саркастичен тон",
