+++ messages/bn.json
@@ -83,7 +83,7 @@
   "settingsTitle": "সেটিংস",
   "settingsOpen": "সেটিংস খুলুন",
   "settingsClose": "সেটিংস বন্ধ করুন",
-  "settingsSectionAppearance": "চেহারা",
+  "settingsSectionGeneral": "সাধারণ",
   "settingsSectionSwitchAccess": "সুইচ অ্যাক্সেস",
   "switchScanningEnabled": "সুইচ স্ক্যানিং চালু করুন",
   "switchScanningMethod": "স্ক্যান পদ্ধতি",
@@ -113,11 +113,11 @@
   "speechVoice": "কণ্ঠস্বর",
   "speechRate": "গতি",
   "speechPitch": "পিচ",
-  "speechVolume": "ভলিউম",
+  "speechVolume": "কথার ভলিউম",
   "speechPreview": "শুনুন",
   "speechVoicePreview": "হ্যালো, এটা আমার কণ্ঠস্বর!",
-  "playbackHighlight": "কথা বলার সময় শব্দগুলি হাইলাইট করুন",
-  "tileSaturation": "রঙের তীব্রতা",
+  "playbackHighlight": "উচ্চারিত শব্দগুলি হাইলাইট করুন",
+  "tileSaturation": "টাইলের রঙের তীব্রতা",
   "tileBorders": "টাইলের সীমানা দেখান",
   "aiCustomInstructions": "কাস্টম নির্দেশনা",
   "aiCustomInstructionsPlaceholder": "যেমন, ব্যঙ্গাত্মক সুরে কথা বলুন",
