+++ messages/cs.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Nastavení",
   "settingsOpen": "Otevřít nastavení",
   "settingsClose": "Zavřít nastavení",
-  "settingsSectionAppearance": "Vzhled",
+  "settingsSectionGeneral": "Obecné",
   "settingsSectionSwitchAccess": "Ovládání spínačem",
   "switchScanningEnabled": "Zapnout skenování spínačem",
   "switchScanningMethod": "Způsob skenování",
@@ -113,11 +113,11 @@
   "speechVoice": "Hlas",
   "speechRate": "Rychlost",
   "speechPitch": "Výška",
-  "speechVolume": "Hlasitost",
+  "speechVolume": "Hlasitost řeči",
   "speechPreview": "Poslechnout",
   "speechVoicePreview": "Ahoj, tohle je můj hlas!",
-  "playbackHighlight": "Zvýraznit slova během mluvení",
-  "tileSaturation": "Intenzita barev",
+  "playbackHighlight": "Zvýraznit mluvená slova",
+  "tileSaturation": "Intenzita barev dlaždic",
   "tileBorders": "Zobrazit ohraničení dlaždic",
   "aiCustomInstructions": "Vlastní pokyny",
   "aiCustomInstructionsPlaceholder": "např. mluv sarkastickým tónem",
