+++ messages/da.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Indstillinger",
   "settingsOpen": "Åbn indstillinger",
   "settingsClose": "Luk indstillinger",
-  "settingsSectionAppearance": "Udseende",
+  "settingsSectionGeneral": "Generelt",
   "settingsSectionSwitchAccess": "Kontakttilgang",
   "switchScanningEnabled": "Slå kontaktscanning til",
   "switchScanningMethod": "Scanningsmetode",
@@ -113,11 +113,11 @@
   "speechVoice": "Stemme",
   "speechRate": "Hastighed",
   "speechPitch": "Tonehøjde",
-  "speechVolume": "Lydstyrke",
+  "speechVolume": "Talelydstyrke",
   "speechPreview": "Lyt",
   "speechVoicePreview": "Hej, det er min stemme!",
-  "playbackHighlight": "Fremhæv ord under tale",
-  "tileSaturation": "Farveintensitet",
+  "playbackHighlight": "Fremhæv talte ord",
+  "tileSaturation": "Felternes farveintensitet",
   "tileBorders": "Vis kant på felter",
   "aiCustomInstructions": "Tilpassede instruktioner",
   "aiCustomInstructionsPlaceholder": "f.eks. tal i en sarkastisk tone",
