+++ messages/de.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Einstellungen",
   "settingsOpen": "Einstellungen öffnen",
   "settingsClose": "Einstellungen schließen",
-  "settingsSectionAppearance": "Erscheinungsbild",
+  "settingsSectionGeneral": "Allgemein",
   "settingsSectionSwitchAccess": "Tastersteuerung",
   "switchScanningEnabled": "Tastenscanning aktivieren",
   "switchScanningMethod": "Scanmethode",
@@ -113,11 +113,11 @@
   "speechVoice": "Stimme",
   "speechRate": "Geschwindigkeit",
   "speechPitch": "Tonhöhe",
-  "speechVolume": "Lautstärke",
+  "speechVolume": "Sprachlautstärke",
   "speechPreview": "Anhören",
   "speechVoicePreview": "Hallo, das ist meine Stimme!",
-  "playbackHighlight": "Wörter beim Sprechen hervorheben",
-  "tileSaturation": "Farbintensität",
+  "playbackHighlight": "Gesprochene Wörter hervorheben",
+  "tileSaturation": "Farbstärke der Kacheln",
   "tileBorders": "Kachel-Rahmen anzeigen",
   "aiCustomInstructions": "Benutzerdefinierte Anweisungen",
   "aiCustomInstructionsPlaceholder": "z. B. in sarkastischem Ton sprechen",
