+++ messages/el.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Ρυθμίσεις",
   "settingsOpen": "Άνοιγμα ρυθμίσεων",
   "settingsClose": "Κλείσιμο ρυθμίσεων",
-  "settingsSectionAppearance": "Εμφάνιση",
+  "settingsSectionGeneral": "Γενικά",
   "settingsSectionSwitchAccess": "Πρόσβαση με διακόπτη",
   "switchScanningEnabled": "Ενεργοποίηση σάρωσης με διακόπτη",
   "switchScanningMethod": "Μέθοδος σάρωσης",
@@ -113,11 +113,11 @@
   "speechVoice": "Φωνή",
   "speechRate": "Ταχύτητα",
   "speechPitch": "Τόνος",
-  "speechVolume": "Ένταση",
+  "speechVolume": "Ένταση ομιλίας",
   "speechPreview": "Ακρόαση",
   "speechVoicePreview": "Γεια, αυτή είναι η φωνή μου!",
-  "playbackHighlight": "Επισήμανση λέξεων κατά την ομιλία",
-  "tileSaturation": "Ένταση χρώματος",
+  "playbackHighlight": "Επισήμανση εκφωνούμενων λέξεων",
+  "tileSaturation": "Ένταση χρώματος πλακιδίων",
   "tileBorders": "Εμφάνιση περιγράμματος πλακιδίων",
   "aiCustomInstructions": "Προσαρμοσμένες οδηγίες",
   "aiCustomInstructionsPlaceholder": "π.χ. μίλα με σαρκαστικό τόνο",
