+++ messages/en.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Settings",
   "settingsOpen": "Open settings",
   "settingsClose": "Close settings",
-  "settingsSectionAppearance": "Appearance",
+  "settingsSectionGeneral": "General",
   "settingsSectionSwitchAccess": "Switch access",
   "switchScanningEnabled": "Enable switch scanning",
   "switchScanningMethod": "Scan method",
@@ -113,11 +113,11 @@
   "speechVoice": "Voice",
   "speechRate": "Rate",
   "speechPitch": "Pitch",
-  "speechVolume": "Volume",
+  "speechVolume": "Speech volume",
   "speechPreview": "Preview",
   "speechVoicePreview": "Hi, this is my voice!",
-  "playbackHighlight": "Highlight words while speaking",
-  "tileSaturation": "Color intensity",
+  "playbackHighlight": "Highlight spoken words",
+  "tileSaturation": "Tile color intensity",
   "tileBorders": "Show tile borders",
   "aiCustomInstructions": "Custom instructions",
   "aiCustomInstructionsPlaceholder": "e.g. speak in a sarcastic tone",
