+++ messages/es.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Ajustes",
   "settingsOpen": "Abrir ajustes",
   "settingsClose": "Cerrar ajustes",
-  "settingsSectionAppearance": "Apariencia",
+  "settingsSectionGeneral": "General",
   "settingsSectionSwitchAccess": "Acceso con pulsador",
   "switchScanningEnabled": "Activar barrido con pulsador",
   "switchScanningMethod": "Método de barrido",
@@ -113,11 +113,11 @@
   "speechVoice": "Voz",
   "speechRate": "Velocidad",
   "speechPitch": "Tono",
-  "speechVolume": "Volumen",
+  "speechVolume": "Volumen de voz",
   "speechPreview": "Escuchar",
   "speechVoicePreview": "¡Hola, esta es mi voz!",
-  "playbackHighlight": "Resaltar palabras mientras se habla",
-  "tileSaturation": "Intensidad del color",
+  "playbackHighlight": "Resaltar palabras pronunciadas",
+  "tileSaturation": "Intensidad de color de los mosaicos",
   "tileBorders": "Mostrar bordes de las casillas",
   "aiCustomInstructions": "Instrucciones personalizadas",
   "aiCustomInstructionsPlaceholder": "p. ej., habla con un tono sarcástico",
