+++ messages/fi.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Asetukset",
   "settingsOpen": "Avaa asetukset",
   "settingsClose": "Sulje asetukset",
-  "settingsSectionAppearance": "Ulkoasu",
+  "settingsSectionGeneral": "Yleiset",
   "settingsSectionSwitchAccess": "Kytkinohjaus",
   "switchScanningEnabled": "Ota kytkinskannaus käyttöön",
   "switchScanningMethod": "Skannaustapa",
@@ -113,11 +113,11 @@
   "speechVoice": "Ääni",
   "speechRate": "Nopeus",
   "speechPitch": "Korkeus",
-  "speechVolume": "Äänenvoimakkuus",
+  "speechVolume": "Puheen äänenvoimakkuus",
   "speechPreview": "Kuuntele",
   "speechVoicePreview": "Hei, tämä on minun ääneni!",
-  "playbackHighlight": "Korosta sanat puheen aikana",
-  "tileSaturation": "Värien voimakkuus",
+  "playbackHighlight": "Korosta puhutut sanat",
+  "tileSaturation": "Ruutujen värien voimakkuus",
   "tileBorders": "Näytä ruutujen reunat",
   "aiCustomInstructions": "Mukautetut ohjeet",
   "aiCustomInstructionsPlaceholder": "esim. puhu sarkastisella sävyllä",
