+++ messages/fr.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Paramètres",
   "settingsOpen": "Ouvrir les paramètres",
   "settingsClose": "Fermer les paramètres",
-  "settingsSectionAppearance": "Apparence",
+  "settingsSectionGeneral": "Général",
   "settingsSectionSwitchAccess": "Accès par contacteur",
   "switchScanningEnabled": "Activer le balayage par contacteur",
   "switchScanningMethod": "Méthode de balayage",
@@ -113,11 +113,11 @@
   "speechVoice": "Voix",
   "speechRate": "Débit",
   "speechPitch": "Hauteur",
-  "speechVolume": "Volume",
+  "speechVolume": "Volume de la parole",
   "speechPreview": "Écouter",
   "speechVoicePreview": "Bonjour, voici ma voix !",
-  "playbackHighlight": "Surligner les mots pendant la lecture à voix haute",
-  "tileSaturation": "Intensité des couleurs",
+  "playbackHighlight": "Surligner les mots prononcés",
+  "tileSaturation": "Intensité des couleurs des tuiles",
   "tileBorders": "Afficher les bordures des cases",
   "aiCustomInstructions": "Instructions personnalisées",
   "aiCustomInstructionsPlaceholder": "p. ex. parler sur un ton sarcastique",
