+++ messages/he.json
@@ -83,7 +83,7 @@
   "settingsTitle": "הגדרות",
   "settingsOpen": "פתיחת הגדרות",
   "settingsClose": "סגירת הגדרות",
-  "settingsSectionAppearance": "מראה",
+  "settingsSectionGeneral": "כללי",
   "settingsSectionSwitchAccess": "גישה באמצעות מתג",
   "switchScanningEnabled": "הפעלת סריקה באמצעות מתג",
   "switchScanningMethod": "שיטת סריקה",
@@ -113,11 +113,11 @@
   "speechVoice": "קול",
   "speechRate": "קצב דיבור",
   "speechPitch": "גובה צליל",
-  "speechVolume": "עוצמת שמע",
+  "speechVolume": "עוצמת דיבור",
   "speechPreview": "השמעה",
   "speechVoicePreview": "שלום, זה הקול שלי!",
-  "playbackHighlight": "הדגשת מילים בעת הדיבור",
-  "tileSaturation": "עוצמת הצבע",
+  "playbackHighlight": "הדגשת מילים מדוברות",
+  "tileSaturation": "עוצמת צבעי המשבצות",
   "tileBorders": "הצגת מסגרות למשבצות",
   "aiCustomInstructions": "הוראות מותאמות אישית",
   "aiCustomInstructionsPlaceholder": "למשל: דבר בטון סרקסטי",
