+++ messages/hi.json
@@ -83,7 +83,7 @@
   "settingsTitle": "सेटिंग्स",
   "settingsOpen": "सेटिंग्स खोलें",
   "settingsClose": "सेटिंग्स बंद करें",
-  "settingsSectionAppearance": "दिखावट",
+  "settingsSectionGeneral": "सामान्य",
   "settingsSectionSwitchAccess": "स्विच एक्सेस",
   "switchScanningEnabled": "स्विच स्कैनिंग चालू करें",
   "switchScanningMethod": "स्कैन विधि",
@@ -113,11 +113,11 @@
   "speechVoice": "आवाज़",
   "speechRate": "गति",
   "speechPitch": "पिच",
-  "speechVolume": "वॉल्यूम",
+  "speechVolume": "वाणी की आवाज़",
   "speechPreview": "सुनें",
   "speechVoicePreview": "नमस्ते, यह मेरी आवाज़ है!",
-  "playbackHighlight": "बोलते समय शब्दों को हाइलाइट करें",
-  "tileSaturation": "रंग की तीव्रता",
+  "playbackHighlight": "बोले गए शब्दों को हाइलाइट करें",
+  "tileSaturation": "टाइल के रंग की तीव्रता",
   "tileBorders": "टाइल की सीमाएँ दिखाएं",
   "aiCustomInstructions": "कस्टम निर्देश",
   "aiCustomInstructionsPlaceholder": "जैसे, व्यंग्यात्मक लहजे में बोलें",
