+++ messages/hr.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Postavke",
   "settingsOpen": "Otvori postavke",
   "settingsClose": "Zatvori postavke",
-  "settingsSectionAppearance": "Izgled",
+  "settingsSectionGeneral": "Općenito",
   "settingsSectionSwitchAccess": "Pristup prekidačem",
   "switchScanningEnabled": "Uključi skeniranje prekidačem",
   "switchScanningMethod": "Način skeniranja",
@@ -113,11 +113,11 @@
   "speechVoice": "Glas",
   "speechRate": "Brzina",
   "speechPitch": "Visina",
-  "speechVolume": "Glasnoća",
+  "speechVolume": "Glasnoća govora",
   "speechPreview": "Poslušaj",
   "speechVoicePreview": "Bok, ovo je moj glas!",
-  "playbackHighlight": "Istakni riječi tijekom govora",
-  "tileSaturation": "Intenzitet boje",
+  "playbackHighlight": "Istakni izgovorene riječi",
+  "tileSaturation": "Intenzitet boje pločica",
   "tileBorders": "Prikaži obrube pločica",
   "aiCustomInstructions": "Prilagođene upute",
   "aiCustomInstructionsPlaceholder": "npr. govori sarkastičnim tonom",
