+++ messages/hu.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Beállítások",
   "settingsOpen": "Beállítások megnyitása",
   "settingsClose": "Beállítások bezárása",
-  "settingsSectionAppearance": "Megjelenés",
+  "settingsSectionGeneral": "Általános",
   "settingsSectionSwitchAccess": "Kapcsolós hozzáférés",
   "switchScanningEnabled": "Kapcsolós pásztázás bekapcsolása",
   "switchScanningMethod": "Pásztázási mód",
@@ -113,11 +113,11 @@
   "speechVoice": "Hang",
   "speechRate": "Sebesség",
   "speechPitch": "Hangmagasság",
-  "speechVolume": "Hangerő",
+  "speechVolume": "Beszéd hangereje",
   "speechPreview": "Meghallgatás",
   "speechVoicePreview": "Szia, ez az én hangom!",
-  "playbackHighlight": "Szavak kiemelése beszéd közben",
-  "tileSaturation": "Színintenzitás",
+  "playbackHighlight": "Kimondott szavak kiemelése",
+  "tileSaturation": "Csempék színintenzitása",
   "tileBorders": "Csempék szegélyének megjelenítése",
   "aiCustomInstructions": "Egyéni utasítások",
   "aiCustomInstructionsPlaceholder": "pl. beszélj szarkasztikus hangnemben",
