+++ messages/id.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Setelan",
   "settingsOpen": "Buka setelan",
   "settingsClose": "Tutup setelan",
-  "settingsSectionAppearance": "Tampilan",
+  "settingsSectionGeneral": "Umum",
   "settingsSectionSwitchAccess": "Akses sakelar",
   "switchScanningEnabled": "Aktifkan pemindaian sakelar",
   "switchScanningMethod": "Metode pemindaian",
@@ -113,11 +113,11 @@
   "speechVoice": "Suara",
   "speechRate": "Kecepatan",
   "speechPitch": "Nada",
-  "speechVolume": "Volume",
+  "speechVolume": "Volume ucapan",
   "speechPreview": "Dengarkan",
   "speechVoicePreview": "Hai, ini suara saya!",
-  "playbackHighlight": "Sorot kata saat berbicara",
-  "tileSaturation": "Intensitas warna",
+  "playbackHighlight": "Sorot kata yang diucapkan",
+  "tileSaturation": "Intensitas warna ubin",
   "tileBorders": "Tampilkan batas petak",
   "aiCustomInstructions": "Instruksi khusus",
   "aiCustomInstructionsPlaceholder": "mis. berbicara dengan nada sarkastis",
