+++ messages/it.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Impostazioni",
   "settingsOpen": "Apri impostazioni",
   "settingsClose": "Chiudi impostazioni",
-  "settingsSectionAppearance": "Aspetto",
+  "settingsSectionGeneral": "Generali",
   "settingsSectionSwitchAccess": "Accesso con sensore",
   "switchScanningEnabled": "Attiva la scansione con sensore",
   "switchScanningMethod": "Metodo di scansione",
@@ -113,11 +113,11 @@
   "speechVoice": "Voce",
   "speechRate": "Velocità",
   "speechPitch": "Tono",
-  "speechVolume": "Volume",
+  "speechVolume": "Volume del parlato",
   "speechPreview": "Ascolta",
   "speechVoicePreview": "Ciao, questa è la mia voce!",
-  "playbackHighlight": "Evidenzia le parole durante il parlato",
-  "tileSaturation": "Intensità del colore",
+  "playbackHighlight": "Evidenzia le parole pronunciate",
+  "tileSaturation": "Intensità del colore delle tessere",
   "tileBorders": "Mostra i bordi delle caselle",
   "aiCustomInstructions": "Istruzioni personalizzate",
   "aiCustomInstructionsPlaceholder": "ad es. parla con tono sarcastico",
