+++ messages/ja.json
@@ -83,7 +83,7 @@
   "settingsTitle": "設定",
   "settingsOpen": "設定を開く",
   "settingsClose": "設定を閉じる",
-  "settingsSectionAppearance": "外観",
+  "settingsSectionGeneral": "一般",
   "settingsSectionSwitchAccess": "スイッチアクセス",
   "switchScanningEnabled": "スイッチスキャンを有効にする",
   "switchScanningMethod": "スキャン方法",
@@ -113,11 +113,11 @@
   "speechVoice": "音声",
   "speechRate": "速度",
   "speechPitch": "ピッチ",
-  "speechVolume": "音量",
+  "speechVolume": "音声の音量",
   "speechPreview": "試聴",
   "speechVoicePreview": "こんにちは、これが私の声です！",
-  "playbackHighlight": "発話中に単語をハイライト表示",
-  "tileSaturation": "色の濃さ",
+  "playbackHighlight": "読み上げ中の単語をハイライト",
+  "tileSaturation": "タイルの色の濃さ",
   "tileBorders": "タイルの枠線を表示",
   "aiCustomInstructions": "カスタム指示",
   "aiCustomInstructionsPlaceholder": "例：皮肉っぽい口調で話す",
