+++ messages/kn.json
@@ -83,7 +83,7 @@
   "settingsTitle": "ಸೆಟ್ಟಿಂಗ್‌ಗಳು",
   "settingsOpen": "ಸೆಟ್ಟಿಂಗ್‌ಗಳನ್ನು ತೆರೆಯಿರಿ",
   "settingsClose": "ಸೆಟ್ಟಿಂಗ್‌ಗಳನ್ನು ಮುಚ್ಚಿ",
-  "settingsSectionAppearance": "ಗೋಚರತೆ",
+  "settingsSectionGeneral": "ಸಾಮಾನ್ಯ",
   "settingsSectionSwitchAccess": "ಸ್ವಿಚ್ ಪ್ರವೇಶ",
   "switchScanningEnabled": "ಸ್ವಿಚ್ ಸ್ಕ್ಯಾನಿಂಗ್ ಸಕ್ರಿಯಗೊಳಿಸಿ",
   "switchScanningMethod": "ಸ್ಕ್ಯಾನ್ ವಿಧಾನ",
@@ -113,11 +113,11 @@
   "speechVoice": "ಧ್ವನಿ",
   "speechRate": "ವೇಗ",
   "speechPitch": "ಪಿಚ್",
-  "speechVolume": "ವಾಲ್ಯೂಮ್",
+  "speechVolume": "ಮಾತಿನ ಧ್ವನಿ ಪ್ರಮಾಣ",
   "speechPreview": "ಕೇಳಿ",
   "speechVoicePreview": "ನಮಸ್ಕಾರ, ಇದು ನನ್ನ ಧ್ವನಿ!",
-  "playbackHighlight": "ಮಾತನಾಡುವಾಗ ಪದಗಳನ್ನು ಹೈಲೈಟ್ ಮಾಡಿ",
-  "tileSaturation": "ಬಣ್ಣದ ತೀವ್ರತೆ",
+  "playbackHighlight": "ಮಾತನಾಡಿದ ಪದಗಳನ್ನು ಹೈಲೈಟ್ ಮಾಡಿ",
+  "tileSaturation": "ಟೈಲ್ ಬಣ್ಣದ ತೀವ್ರತೆ",
   "tileBorders": "ಟೈಲ್ ಗಡಿಗಳನ್ನು ತೋರಿಸಿ",
   "aiCustomInstructions": "ಕಸ್ಟಮ್ ಸೂಚನೆಗಳು",
   "aiCustomInstructionsPlaceholder": "ಉದಾ., ವ್ಯಂಗ್ಯ ಧ್ವನಿಯಲ್ಲಿ ಮಾತನಾಡಿ",
