+++ messages/ko.json
@@ -83,7 +83,7 @@
   "settingsTitle": "설정",
   "settingsOpen": "설정 열기",
   "settingsClose": "설정 닫기",
-  "settingsSectionAppearance": "모양",
+  "settingsSectionGeneral": "일반",
   "settingsSectionSwitchAccess": "스위치 제어",
   "switchScanningEnabled": "스위치 스캔 사용",
   "switchScanningMethod": "스캔 방식",
@@ -113,11 +113,11 @@
   "speechVoice": "음성",
   "speechRate": "속도",
   "speechPitch": "음높이",
-  "speechVolume": "음량",
+  "speechVolume": "음성 볼륨",
   "speechPreview": "미리 듣기",
   "speechVoicePreview": "안녕하세요, 제 목소리예요!",
-  "playbackHighlight": "말할 때 단어 강조 표시",
-  "tileSaturation": "색상 강도",
+  "playbackHighlight": "말한 단어 강조 표시",
+  "tileSaturation": "타일 색상 강도",
   "tileBorders": "타일 테두리 표시",
   "aiCustomInstructions": "맞춤 지침",
   "aiCustomInstructionsPlaceholder": "예: 냉소적인 어조로 말하기",
