+++ messages/nl.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Instellingen",
   "settingsOpen": "Instellingen openen",
   "settingsClose": "Instellingen sluiten",
-  "settingsSectionAppearance": "Uiterlijk",
+  "settingsSectionGeneral": "Algemeen",
   "settingsSectionSwitchAccess": "Schakelaarbediening",
   "switchScanningEnabled": "Scannen met schakelaar inschakelen",
   "switchScanningMethod": "Scanmethode",
@@ -113,11 +113,11 @@
   "speechVoice": "Stem",
   "speechRate": "Snelheid",
   "speechPitch": "Toonhoogte",
-  "speechVolume": "Volume",
+  "speechVolume": "Spraakvolume",
   "speechPreview": "Beluisteren",
   "speechVoicePreview": "Hoi, dit is mijn stem!",
-  "playbackHighlight": "Woorden markeren tijdens het spreken",
-  "tileSaturation": "Kleurintensiteit",
+  "playbackHighlight": "Gesproken woorden markeren",
+  "tileSaturation": "Kleurintensiteit van tegels",
   "tileBorders": "Tegelranden weergeven",
   "aiCustomInstructions": "Aangepaste instructies",
   "aiCustomInstructionsPlaceholder": "bijv. spreek op een sarcastische toon",
