+++ messages/pl.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Ustawienia",
   "settingsOpen": "Otwórz ustawienia",
   "settingsClose": "Zamknij ustawienia",
-  "settingsSectionAppearance": "Wygląd",
+  "settingsSectionGeneral": "Ogólne",
   "settingsSectionSwitchAccess": "Dostęp przełącznikiem",
   "switchScanningEnabled": "Włącz skanowanie przełącznikiem",
   "switchScanningMethod": "Metoda skanowania",
@@ -113,11 +113,11 @@
   "speechVoice": "Głos",
   "speechRate": "Tempo",
   "speechPitch": "Wysokość",
-  "speechVolume": "Głośność",
+  "speechVolume": "Głośność mowy",
   "speechPreview": "Posłuchaj",
   "speechVoicePreview": "Cześć, to mój głos!",
-  "playbackHighlight": "Podświetlaj słowa podczas mówienia",
-  "tileSaturation": "Intensywność kolorów",
+  "playbackHighlight": "Podświetlaj wypowiadane słowa",
+  "tileSaturation": "Intensywność kolorów kafelków",
   "tileBorders": "Pokaż obramowania kafelków",
   "aiCustomInstructions": "Instrukcje niestandardowe",
   "aiCustomInstructionsPlaceholder": "np. mów sarkastycznym tonem",
