+++ messages/pt.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Configurações",
   "settingsOpen": "Abrir configurações",
   "settingsClose": "Fechar configurações",
-  "settingsSectionAppearance": "Aparência",
+  "settingsSectionGeneral": "Geral",
   "settingsSectionSwitchAccess": "Acesso por acionador",
   "switchScanningEnabled": "Ativar varredura por acionador",
   "switchScanningMethod": "Método de varredura",
@@ -113,11 +113,11 @@
   "speechVoice": "Voz",
   "speechRate": "Velocidade",
   "speechPitch": "Tom",
-  "speechVolume": "Volume",
+  "speechVolume": "Volume da fala",
   "speechPreview": "Ouvir",
   "speechVoicePreview": "Olá, esta é a minha voz!",
-  "playbackHighlight": "Destacar palavras durante a fala",
-  "tileSaturation": "Intensidade da cor",
+  "playbackHighlight": "Destacar palavras faladas",
+  "tileSaturation": "Intensidade da cor dos blocos",
   "tileBorders": "Mostrar contornos dos blocos",
   "aiCustomInstructions": "Instruções personalizadas",
   "aiCustomInstructionsPlaceholder": "ex.: fale com um tom sarcástico",
