+++ messages/ro.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Setări",
   "settingsOpen": "Deschide setările",
   "settingsClose": "Închide setările",
-  "settingsSectionAppearance": "Aspect",
+  "settingsSectionGeneral": "Generale",
   "settingsSectionSwitchAccess": "Acces cu comutator",
   "switchScanningEnabled": "Activează scanarea cu comutator",
   "switchScanningMethod": "Metodă de scanare",
@@ -113,11 +113,11 @@
   "speechVoice": "Voce",
   "speechRate": "Viteză",
   "speechPitch": "Înălțime",
-  "speechVolume": "Volum",
+  "speechVolume": "Volumul vorbirii",
   "speechPreview": "Ascultă",
   "speechVoicePreview": "Bună, aceasta este vocea mea!",
-  "playbackHighlight": "Evidențiază cuvintele în timpul vorbirii",
-  "tileSaturation": "Intensitatea culorii",
+  "playbackHighlight": "Evidențiază cuvintele rostite",
+  "tileSaturation": "Intensitatea culorii dalelor",
   "tileBorders": "Afișează marginile casetelor",
   "aiCustomInstructions": "Instrucțiuni personalizate",
   "aiCustomInstructionsPlaceholder": "ex. vorbește pe un ton sarcastic",
