+++ messages/ru.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Настройки",
   "settingsOpen": "Открыть настройки",
   "settingsClose": "Закрыть настройки",
-  "settingsSectionAppearance": "Внешний вид",
+  "settingsSectionGeneral": "Общие",
   "settingsSectionSwitchAccess": "Управление переключателем",
   "switchScanningEnabled": "Включить сканирование переключателем",
   "switchScanningMethod": "Метод сканирования",
@@ -113,11 +113,11 @@
   "speechVoice": "Голос",
   "speechRate": "Скорость",
   "speechPitch": "Высота тона",
-  "speechVolume": "Громкость",
+  "speechVolume": "Громкость речи",
   "speechPreview": "Прослушать",
   "speechVoicePreview": "Привет, это мой голос!",
-  "playbackHighlight": "Подсвечивать слова во время произношения",
-  "tileSaturation": "Насыщенность цвета",
+  "playbackHighlight": "Подсвечивать произносимые слова",
+  "tileSaturation": "Насыщенность цвета плиток",
   "tileBorders": "Показывать границы плиток",
   "aiCustomInstructions": "Пользовательские инструкции",
   "aiCustomInstructionsPlaceholder": "напр. говори саркастичным тоном",
