+++ messages/sk.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Nastavenia",
   "settingsOpen": "Otvoriť nastavenia",
   "settingsClose": "Zavrieť nastavenia",
-  "settingsSectionAppearance": "Vzhľad",
+  "settingsSectionGeneral": "Všeobecné",
   "settingsSectionSwitchAccess": "Ovládanie spínačom",
   "switchScanningEnabled": "Zapnúť skenovanie spínačom",
   "switchScanningMethod": "Spôsob skenovania",
@@ -113,11 +113,11 @@
   "speechVoice": "Hlas",
   "speechRate": "Rýchlosť",
   "speechPitch": "Výška",
-  "speechVolume": "Hlasitosť",
+  "speechVolume": "Hlasitosť reči",
   "speechPreview": "Vypočuť",
   "speechVoicePreview": "Ahoj, toto je môj hlas!",
-  "playbackHighlight": "Zvýrazniť slová počas hovorenia",
-  "tileSaturation": "Intenzita farieb",
+  "playbackHighlight": "Zvýrazniť hovorené slová",
+  "tileSaturation": "Intenzita farieb dlaždíc",
   "tileBorders": "Zobraziť okraje dlaždíc",
   "aiCustomInstructions": "Vlastné pokyny",
   "aiCustomInstructionsPlaceholder": "napr. hovor sarkastickým tónom",
