+++ messages/sl.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Nastavitve",
   "settingsOpen": "Odpri nastavitve",
   "settingsClose": "Zapri nastavitve",
-  "settingsSectionAppearance": "Videz",
+  "settingsSectionGeneral": "Splošno",
   "settingsSectionSwitchAccess": "Dostop s stikalom",
   "switchScanningEnabled": "Vklopi skeniranje s stikalom",
   "switchScanningMethod": "Način skeniranja",
@@ -113,11 +113,11 @@
   "speechVoice": "Glas",
   "speechRate": "Hitrost",
   "speechPitch": "Višina",
-  "speechVolume": "Glasnost",
+  "speechVolume": "Glasnost govora",
   "speechPreview": "Poslušaj",
   "speechVoicePreview": "Živjo, to je moj glas!",
-  "playbackHighlight": "Označi besede med govorjenjem",
-  "tileSaturation": "Intenzivnost barve",
+  "playbackHighlight": "Označi izgovorjene besede",
+  "tileSaturation": "Intenzivnost barve ploščic",
   "tileBorders": "Prikaži obrobe ploščic",
   "aiCustomInstructions": "Navodila po meri",
   "aiCustomInstructionsPlaceholder": "npr. govori sarkastičnim tonom",
