+++ messages/sv.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Inställningar",
   "settingsOpen": "Öppna inställningar",
   "settingsClose": "Stäng inställningar",
-  "settingsSectionAppearance": "Utseende",
+  "settingsSectionGeneral": "Allmänt",
   "settingsSectionSwitchAccess": "Kontaktstyrning",
   "switchScanningEnabled": "Aktivera kontaktskanning",
   "switchScanningMethod": "Skanningsmetod",
@@ -113,11 +113,11 @@
   "speechVoice": "Röst",
   "speechRate": "Hastighet",
   "speechPitch": "Tonhöjd",
-  "speechVolume": "Volym",
+  "speechVolume": "Talvolym",
   "speechPreview": "Lyssna",
   "speechVoicePreview": "Hej, det här är min röst!",
-  "playbackHighlight": "Markera ord medan de talas",
-  "tileSaturation": "Färgintensitet",
+  "playbackHighlight": "Markera talade ord",
+  "tileSaturation": "Färgintensitet för rutor",
   "tileBorders": "Visa kantlinjer på rutor",
   "aiCustomInstructions": "Anpassade instruktioner",
   "aiCustomInstructionsPlaceholder": "t.ex. tala med sarkastisk ton",
