+++ messages/ta.json
@@ -83,7 +83,7 @@
   "settingsTitle": "அமைப்புகள்",
   "settingsOpen": "அமைப்புகளைத் திற",
   "settingsClose": "அமைப்புகளை மூடு",
-  "settingsSectionAppearance": "தோற்றம்",
+  "settingsSectionGeneral": "பொது",
   "settingsSectionSwitchAccess": "சுவிட்ச் அணுகல்",
   "switchScanningEnabled": "சுவிட்ச் ஸ்கேனிங்கை இயக்கு",
   "switchScanningMethod": "ஸ்கேன் முறை",
@@ -113,11 +113,11 @@
   "speechVoice": "குரல்",
   "speechRate": "வேகம்",
   "speechPitch": "சுருதி",
-  "speechVolume": "ஒலியளவு",
+  "speechVolume": "பேச்சு ஒலியளவு",
   "speechPreview": "கேள்",
   "speechVoicePreview": "வணக்கம், இது என் குரல்!",
-  "playbackHighlight": "பேசும்போது சொற்களை தனிப்படுத்து",
-  "tileSaturation": "வண்ணத் தீவிரம்",
+  "playbackHighlight": "பேசப்படும் சொற்களைத் தனிப்படுத்து",
+  "tileSaturation": "கட்டங்களின் வண்ணத் தீவிரம்",
   "tileBorders": "டைல் எல்லைகளைக் காட்டு",
   "aiCustomInstructions": "தனிப்பயன் வழிமுறைகள்",
   "aiCustomInstructionsPlaceholder": "எ.கா., கிண்டலான தொனியில் பேசு",
