+++ messages/te.json
@@ -83,7 +83,7 @@
   "settingsTitle": "సెట్టింగ్‌లు",
   "settingsOpen": "సెట్టింగ్‌లను తెరువు",
   "settingsClose": "సెట్టింగ్‌లను మూసివేయి",
-  "settingsSectionAppearance": "రూపం",
+  "settingsSectionGeneral": "సాధారణ",
   "settingsSectionSwitchAccess": "స్విచ్ యాక్సెస్",
   "switchScanningEnabled": "స్విచ్ స్కానింగ్‌ను ప్రారంభించండి",
   "switchScanningMethod": "స్కాన్ పద్ధతి",
@@ -113,11 +113,11 @@
   "speechVoice": "వాయిస్",
   "speechRate": "వేగం",
   "speechPitch": "పిచ్",
-  "speechVolume": "వాల్యూమ్",
+  "speechVolume": "మాటల వాల్యూమ్",
   "speechPreview": "విను",
   "speechVoicePreview": "హాయ్, ఇది నా వాయిస్!",
-  "playbackHighlight": "మాట్లాడేటప్పుడు పదాలను హైలైట్ చేయి",
-  "tileSaturation": "రంగు తీవ్రత",
+  "playbackHighlight": "పలికిన పదాలను హైలైట్ చేయి",
+  "tileSaturation": "టైల్ రంగు తీవ్రత",
   "tileBorders": "టైల్ సరిహద్దులను చూపించు",
   "aiCustomInstructions": "అనుకూల సూచనలు",
   "aiCustomInstructionsPlaceholder": "ఉదా., వ్యంగ్యమైన స్వరంతో మాట్లాడు",
