+++ messages/th.json
@@ -83,7 +83,7 @@
   "settingsTitle": "การตั้งค่า",
   "settingsOpen": "เปิดการตั้งค่า",
   "settingsClose": "ปิดการตั้งค่า",
-  "settingsSectionAppearance": "รูปลักษณ์",
+  "settingsSectionGeneral": "ทั่วไป",
   "settingsSectionSwitchAccess": "การเข้าถึงด้วยสวิตช์",
   "switchScanningEnabled": "เปิดใช้การสแกนด้วยสวิตช์",
   "switchScanningMethod": "วิธีการสแกน",
@@ -113,11 +113,11 @@
   "speechVoice": "เสียง",
   "speechRate": "ความเร็ว",
   "speechPitch": "ระดับเสียงสูงต่ำ",
-  "speechVolume": "ความดัง",
+  "speechVolume": "ระดับเสียงพูด",
   "speechPreview": "ฟัง",
   "speechVoicePreview": "สวัสดี นี่คือเสียงของฉัน!",
-  "playbackHighlight": "ไฮไลต์คำขณะพูด",
-  "tileSaturation": "ความเข้มของสี",
+  "playbackHighlight": "ไฮไลต์คำที่พูด",
+  "tileSaturation": "ความเข้มของสีไทล์",
   "tileBorders": "แสดงขอบไทล์",
   "aiCustomInstructions": "คำสั่งที่กำหนดเอง",
   "aiCustomInstructionsPlaceholder": "เช่น พูดด้วยน้ำเสียงประชดประชัน",
