+++ messages/tr.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Ayarlar",
   "settingsOpen": "Ayarları aç",
   "settingsClose": "Ayarları kapat",
-  "settingsSectionAppearance": "Görünüm",
+  "settingsSectionGeneral": "Genel",
   "settingsSectionSwitchAccess": "Anahtar erişimi",
   "switchScanningEnabled": "Anahtarla taramayı etkinleştir",
   "switchScanningMethod": "Tarama yöntemi",
@@ -113,11 +113,11 @@
   "speechVoice": "Ses",
   "speechRate": "Hız",
   "speechPitch": "Perde",
-  "speechVolume": "Ses düzeyi",
+  "speechVolume": "Konuşma ses düzeyi",
   "speechPreview": "Dinle",
   "speechVoicePreview": "Merhaba, bu benim sesim!",
-  "playbackHighlight": "Konuşurken kelimeleri vurgula",
-  "tileSaturation": "Renk yoğunluğu",
+  "playbackHighlight": "Söylenen kelimeleri vurgula",
+  "tileSaturation": "Karo renk yoğunluğu",
   "tileBorders": "Kare kenarlıklarını göster",
   "aiCustomInstructions": "Özel talimatlar",
   "aiCustomInstructionsPlaceholder": "ör. alaycı bir tonda konuş",
