+++ messages/uk.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Налаштування",
   "settingsOpen": "Відкрити налаштування",
   "settingsClose": "Закрити налаштування",
-  "settingsSectionAppearance": "Зовнішній вигляд",
+  "settingsSectionGeneral": "Загальні",
   "settingsSectionSwitchAccess": "Керування перемикачем",
   "switchScanningEnabled": "Увімкнути сканування перемикачем",
   "switchScanningMethod": "Метод сканування",
@@ -113,11 +113,11 @@
   "speechVoice": "Голос",
   "speechRate": "Швидкість",
   "speechPitch": "Висота тону",
-  "speechVolume": "Гучність",
+  "speechVolume": "Гучність мовлення",
   "speechPreview": "Прослухати",
   "speechVoicePreview": "Привіт, це мій голос!",
-  "playbackHighlight": "Підсвічувати слова під час озвучення",
-  "tileSaturation": "Насиченість кольору",
+  "playbackHighlight": "Підсвічувати вимовлені слова",
+  "tileSaturation": "Насиченість кольору плиток",
   "tileBorders": "Показувати межі плиток",
   "aiCustomInstructions": "Власні інструкції",
   "aiCustomInstructionsPlaceholder": "напр. говори саркастичним тоном",
