+++ messages/vi.json
@@ -83,7 +83,7 @@
   "settingsTitle": "Cài đặt",
   "settingsOpen": "Mở cài đặt",
   "settingsClose": "Đóng cài đặt",
-  "settingsSectionAppearance": "Giao diện",
+  "settingsSectionGeneral": "Chung",
   "settingsSectionSwitchAccess": "Truy cập bằng công tắc",
   "switchScanningEnabled": "Bật quét bằng công tắc",
   "switchScanningMethod": "Phương thức quét",
@@ -113,11 +113,11 @@
   "speechVoice": "Giọng nói",
   "speechRate": "Tốc độ",
   "speechPitch": "Cao độ",
-  "speechVolume": "Âm lượng",
+  "speechVolume": "Âm lượng giọng nói",
   "speechPreview": "Nghe thử",
   "speechVoicePreview": "Xin chào, đây là giọng nói của tôi!",
-  "playbackHighlight": "Đánh dấu từ khi nói",
-  "tileSaturation": "Cường độ màu",
+  "playbackHighlight": "Đánh dấu các từ được nói",
+  "tileSaturation": "Cường độ màu ô",
   "tileBorders": "Hiển thị viền ô",
   "aiCustomInstructions": "Hướng dẫn tùy chỉnh",
   "aiCustomInstructionsPlaceholder": "ví dụ: nói với giọng điệu châm biếm",
