+++ messages/zh.json
@@ -83,7 +83,7 @@
   "settingsTitle": "设置",
   "settingsOpen": "打开设置",
   "settingsClose": "关闭设置",
-  "settingsSectionAppearance": "外观",
+  "settingsSectionGeneral": "常规",
   "settingsSectionSwitchAccess": "开关控制",
   "switchScanningEnabled": "启用开关扫描",
   "switchScanningMethod": "扫描方式",
@@ -113,11 +113,11 @@
   "speechVoice": "语音",
   "speechRate": "语速",
   "speechPitch": "音调",
-  "speechVolume": "音量",
+  "speechVolume": "语音音量",
   "speechPreview": "试听",
   "speechVoicePreview": "你好，这是我的声音！",
-  "playbackHighlight": "说话时高亮显示单词",
-  "tileSaturation": "颜色强度",
+  "playbackHighlight": "高亮显示朗读的词语",
+  "tileSaturation": "图块颜色强度",
   "tileBorders": "显示图块边框",
   "aiCustomInstructions": "自定义指令",
   "aiCustomInstructionsPlaceholder": "例如：以讽刺的语气说话",
