+++ src/app/settings/board-settings.test.tsx
@@ -1,32 +1,20 @@
 import { AppProviders } from "@shared/providers/app-providers";
 import { expectNoA11yViolations } from "@shared/testing/axe";
-import { beforeEach, describe, expect, test, vi } from "vitest";
+import { describe, expect, test, vi } from "vitest";
 import { userEvent } from "vitest/browser";
 import { render } from "vitest-browser-react";
 import { BoardSettings } from "./board-settings";
 
 describe("BoardSettings", () => {
-  beforeEach(() => {
-    localStorage.clear();
-  });
-
-  test("toggles the highlight-while-speaking switch with no a11y violations", async () => {
+  test("renders the board controls with no a11y violations", async () => {
     const screen = await render(
       <AppProviders>
         <BoardSettings />
       </AppProviders>,
     );
 
-    const highlightSwitch = screen.getByRole("switch", {
-      name: "Highlight words while speaking",
-    });
-
-    await expect.element(highlightSwitch).not.toBeChecked();
-    await highlightSwitch.click();
-    await expect.element(highlightSwitch).toBeChecked();
-
     const saturationSlider = screen.getByRole("slider", {
-      name: "Color intensity",
+      name: "Tile color intensity",
     });
 
     await expect.element(saturationSlider).toBeVisible();
@@ -74,7 +62,7 @@ describe("BoardSettings", () => {
     );
 
     const saturationSlider = screen.getByRole("slider", {
-      name: "Color intensity",
+      name: "Tile color intensity",
     });
 
     saturationSlider.element().focus();
@@ -83,7 +71,7 @@ describe("BoardSettings", () => {
     const valueNow = Number(
       saturationSlider.element().getAttribute("aria-valuenow"),
     );
-    expect(valueNow).toBeLessThan(1);
+    expect(valueNow).toBe(0.9);
 
     await vi.waitFor(() => {
       const stored = localStorage.getItem("tile-color-config");
